Готовая корзина

+++ app/Http/Controllers/BasketController.php
@@ -16,19 +16,44 @@ class BasketController extends Controller
         return view('basket', compact('order'));
     }
 
+    public function basketConfirm(Request $request)
+    {
+        $orderId = session('orderId');
+        if (is_null($orderId)) {
+            return redirect(route('main'));
+        }
+        $order = Order::find($orderId);
+        $success = $order->saveOrder($request->title, $request->name, $request->phone, $request->address);
+
+        if ($success) {
+            session()->flash('success', 'Ваш заказ принят!');
+        } else {
+            session()->flash('failure', 'Ошибка');
+        }
+
+
+        return redirect(route('main'));
+    }
+
     public function basketAdd($productId)
     {
         $orderId = session('orderId');
         if (is_null($orderId)) {
             $order = Order::create();
-            dump($order);
             session(['orderId' => $order->id]);
         } else {
             $order = Order::find($orderId);
         }
-        $order->products()->attach($productId);
 
-        return redirect()->route('main');
+        if ($order->products->contains($productId)) {
+            $pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+            $pivotRow->count++;
+            $pivotRow->update();
+        } else {
+            $order->products()->attach($productId);
+        }
+
+        return redirect()->route('basket');
     }
 
     public function basketRemove($productId)
@@ -36,8 +61,16 @@ class BasketController extends Controller
         $orderId = session('orderId');
         if (!is_null($orderId)) {
             $order = Order::find($orderId);
-            $order->products()->detach($productId);
+            if ($order->products->contains($productId)) {
+                $pivotRow = $order->products()->where('product_id', $productId)->first()->pivot;
+                if ($pivotRow->count < 2) {
+                    $order->products()->detach($productId);
+                } else {
+                    $pivotRow->count--;
+                    $pivotRow->update();
+                }
+            }
+            return redirect()->route('basket');
         }
-        return redirect()->route('basket');
     }
 }
